Refactor 'BuildId' to 'BuildNumber'
