+++ src/lib/components/BuildNumber.svelte
